Bundle LuigiClient with Babel
