+++ client/package.json
@@ -2,8 +2,10 @@
   "name": "@kyma-project/luigi-client",
   "version": "0.4.3",
   "description": "Javascript library supporting consumers of the Luigi framework",
-  "main": "luigi-client.js",
+  "main": "public/luigi-client.bundle.js",
   "scripts": {
+    "bundle": "webpack --mode production",
+    "bundle-develop": "webpack --mode development --watch",
     "test": "echo \"Error: no test specified\" && exit 1",
     "docu": "npm run docu:validate && npm run docu:generate:section",
     "docu:generate:new-file": "documentation build luigi-client.js -f md --markdown-toc=false -o ../docs/luigi-client-api.md",
@@ -25,6 +27,11 @@
   ],
   "license": "Apache-2.0",
   "devDependencies": {
-    "documentation": "^8.1.2"
+    "@babel/core": "^7.2.2",
+    "@babel/preset-env": "^7.2.3",
+    "babel-loader": "^8.0.5",
+    "documentation": "^8.1.2",
+    "webpack": "^4.28.1",
+    "webpack-cli": "^3.2.1"
   }
 }
