+++ core/examples/luigi-sample-angular/src/assets/basicexternal.html
@@ -20,7 +20,7 @@
   a:hover {
     color: #085caf;
   }
-  </style>
+</style>
 </head>
 <body>
 <a name="ext"></a>
@@ -51,6 +51,6 @@
   </div>
 </section>
 <div class="section-spacer"></div>
-<script type="text/javascript" src="/luigi-client/luigi-client.js"></script>
+<script type="text/javascript" src="/luigi-client/public/luigi-client.bundle.js"></script>
 </body>
 </html>
