+++ core/examples/luigi-sample-angular/src/assets/sampleexternal.html
@@ -20,7 +20,7 @@
   a:hover {
     color: #085caf;
   }
-  </style>
+</style>
 </head>
 <body>
 <a name="ext"></a>
@@ -62,7 +62,7 @@ LuigiClient.uxManager().hideLoadingIndicator();</pre>
   </div>
 </section>
 <div class="section-spacer"></div>
-<script type="text/javascript" src="/luigi-client/luigi-client.js"></script>
+<script type="text/javascript" src="/luigi-client/public/luigi-client.bundle.js"></script>
 <script type="text/javascript">
 // If node has loadingIndicator.hideAutomatically set to false, we need to manually
 // trigger LuigiClient.uxManager().hideLoadingIndicator();
